requirements.txt
streamlit==1.30.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
pytz==2023.3 